Rename file formats to formats
