+++ docs/documentation/formats.md
@@ -1,4 +1,4 @@
-# File formats
+# Formats
 
 Since version 1.2.0 of Datafaker it's possible to export generated data to a file format of your choice.
 
@@ -6,6 +6,7 @@ The currently support file formats are:
 
 * CSV
 * JSON
+* SQL
 * YAML
 * XML
