+++ mkdocs.yml
@@ -100,7 +100,7 @@ nav:
       - Collections: documentation/collections.md
       - Date format: documentation/date-format.md
       - Expressions: documentation/expressions.md
-      - File formats: documentation/file-formats.md
+      - Formats: documentation/formats.md
       - Unique values: documentation/unique-values.md
     - Performance: documentation/performance.md
     - Javadoc: 'https://s01.oss.sonatype.org/service/local/repositories/releases/archive/net/datafaker/datafaker/1.6.0/datafaker-1.6.0-javadoc.jar/!/index.html'
