+++ src/main/java/net/datafaker/formats/Csv.java
@@ -1,4 +1,4 @@
-package net.datafaker.fileformats;
+package net.datafaker.formats;
 
 import net.datafaker.FakeCollection;
