+++ src/main/java/net/datafaker/formats/Xml.java
@@ -1,4 +1,4 @@
-package net.datafaker.fileformats;
+package net.datafaker.formats;
 
 import java.util.Collection;
 import java.util.Collections;
