+++ src/main/java/net/datafaker/providers/base/BaseFaker.java
@@ -1,7 +1,7 @@
 package net.datafaker.providers.base;
 
 import net.datafaker.FakeCollection;
-import net.datafaker.fileformats.Json;
+import net.datafaker.formats.Json;
 import net.datafaker.service.FakeValuesService;
 import net.datafaker.service.FakerContext;
 import net.datafaker.service.RandomService;
