+++ src/test/java/net/datafaker/formats/CsvTest.java
@@ -1,4 +1,4 @@
-package net.datafaker.fileformats;
+package net.datafaker.formats;
 
 import net.datafaker.AbstractFakerTest;
 import net.datafaker.providers.base.BaseFaker;
