+++ src/test/java/net/datafaker/formats/SqlTest.java
@@ -1,4 +1,4 @@
-package net.datafaker.fileformats;
+package net.datafaker.formats;
 
 import net.datafaker.transformations.Schema;
 import net.datafaker.transformations.SqlTransformer;
