+++ src/test/java/net/datafaker/formats/XmlTest.java
@@ -1,4 +1,4 @@
-package net.datafaker.fileformats;
+package net.datafaker.formats;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
